+++ test/transformer/transformer_maps_test.dart
@@ -10,3 +10,11 @@ import "test_models_for_maps_serialization_rules.dart";
 import "package:unittest/unittest.dart";
 
 part "transformer_test_core.dart";
+
+formatSpecificTests(serialization1, serialization2) {
+  test("Verify that we are actually writing in map format", () {
+    var written = serialization1.write(thing1);
+    expect(written is Map, isTrue);
+    expect(written['name'], 'thing1');
+  });
+}