+++ test/transformer/transformer_test.dart
@@ -10,3 +10,12 @@ import "test_models_serialization_rules.dart";
 import "package:unittest/unittest.dart";
 
 part "transformer_test_core.dart";
+
+formatSpecificTests(serialization1, serialization2) {
+  test("Verify that we are actually writing in simple JSON format as a list",
+      () {
+    var written = serialization1.write(thing1);
+    expect(written is List, isTrue);
+    expect(written[1], 'thing1');
+  });
+}